+++ spec/app/models/sound.rb
@@ -1,0 +1,5 @@
+class Sound
+  include Mongoid::Document
+  field :active, type: Boolean
+  default_scope where(active: true)
+end